File loader minor fixes

+++ src/Specialist.cpp
@@ -645,7 +645,7 @@ bool SpecFileLoader::loadFile(const std::string& fileName, bool run)
         m_platform->reset();
         Cpu8080Compatible* cpu = dynamic_cast<Cpu8080Compatible*>(m_platform->getCpu());
         if (cpu) {
-            g_emulation->exec(cpu->getKDiv() * 2000000);
+            g_emulation->exec(int64_t(cpu->getKDiv()) * 2000000);
             cpu->setPC(begAddr);
         }
     }
